web-flask/requirements.txt
# pip install -r requirements.txt
flask
flask-dance[sqla]
blinker
python-dotenv
requests